[PATCH] fix cleanup using VACUUM ANALYZE (RT#149134)
Running VACUUM ANALYZE has never been a proper way to both VACUUM and
ANALYZE. The term after VACUUM is meant to be the schema to work on.
SQLite versions prior to 3.15 (DBD::SQLite 1.52) would ignore the
schema and just operate on the main schema.

Fix the two cleanup tests and the documentation to use "VACUUM;
ANALYZE", which will properly run both operations.